[MOREL-244] Simplify syntax of `order` step, and remove the `desc` keyword
In this change, we change the syntax of the `order` step to
take a single value (provided it is a type that Morel is able
to compare, per #282). Its syntax changes from

```
  step -> ...
    | order orderItem , ... , orderItem

  orderItem -> exp [ desc ]
```
to
```
  step -> ...
    | order exp
```

Warn if the argument to `order` is a record whose fields do
not appear in alphabetical order.

Fixes #244