[MOREL-288] Improve syntax of `group` and `compute` steps
This is a breaking change, but the syntax is more concise and powerful.

What comes after the `group` and `compute` keywords is now
syntactically a single expression. If you have multiple group
fields or multiple compute expressions you should use a
record constructor, so that the fields are assigned names.
The fields in the output are the union of the names in the
group key and compute clause.

The output is an atom if there is one field (i.e. one group
key and no compute expressions or an empty group key and one
compute expression) and the non-empty expression does not use
record-constructor syntax. For example `from e in scott.emps
group e.deptno` and `from e in scott.emps group {} compute
count over ()` are both atomic, and have type `int bag`.

Inside `compute` can now be full expressions. The expressions
inside and outside the aggregate function can now be
non-trivial. For example, `1 + sum over e.sal * 2` is
equivalent to SQL `1 + SUM(e.sal)`. A `compute` expression
can have multiple aggregate functions (e.g. `((min over
e.sal) + (max over e.sal)) / 2`) or no aggregate functions
(e.g. `1`).

Because in Morel *any* expression can be an aggregate
function (it just needs to be a function that takes a list
or a bag as its argument) we need a syntactic way to
separate an aggregate function from its argument. Previously
that was the `of` keyword, but now this is the `over` keyword.

The precedence of 'over' is 7.5 on the left, 5 on the right.
This may seem a bit weird, but allows you to write `1 + sum
over j * 2` without parentheses. Maybe we'll adjust precedence
later.

An `over` expression is an expression, but it is an error if
`over` occurs outside of `compute`, or if `over` occurs
within the argument of `over`. But it is OK if the compute
expression is a query, such as `from`, and if `over` occurs
within that query.

Update the documentation in `query.md` and `reference.md` for
the new `group` and `compute` syntax. Also modify examples
for the previous change from `list` to `bag`. Add the queries
in `query.md` into the `blog.smli` test, to keep us honest.

Fixes #288