[MOREL-285] Type abbreviations, also known as alias types, declared using the `type` keyword
For example, `type myInt = int` declares type `myInt` as
an alias for `int`.

There are a couple of limitations in the current
implementation. First, alias types may not have type
variables (e.g. `type 'a collection = 'a list`).

Second, the type of a value binding can only be deduced to be
an alias type in simple cases. This is because core type
derivation (using the Hindley-Milner algorithm) operates on
expanded types. When type derivation has completed
successfully, we then try to convert the expanded type to an
alias type. We can currently do this only if the pattern is
annotated with an alias type (e.g. `val i: myInt = 5`) or if
the expression is annotated with an alias type (e.g. `val j
= 5: myInt`), or the first element of a list (e.g. `val l =
[7: myInt]`). More cases could be added in the future.

Fixes #285