Assert that in an assignment `val p = e`, pattern and expression have same type
This paves the way for #285 (alias types, and the the `type`
keyword), where we wish the pattern to have the richer type
and the expression to have the expanded type. For example,
in `val i: myInt = 5: int` the pattern has type `myInt` and
the expression has type `int`. Alias typeswill be the
exception to the rule, and this change establishes the rule.

We fix a few bugs:
 * We were incorrectly inferring the type of records with
   `with`, such as `{r with {b = "string" option}`, when
   the type of `r.b` had a type variable such as `alpha
   option`.
 * When expanding unconstrained variables to extents, we
   were treating the extents as having `list` type, whereas
   type inference had given them `bag` type.