Aggregate functions should adapt to the collection type of the input (#271)
In the `compute` clause of a query, you can use the infix `over`
operator to apply aggregate functions to the input or a projection of
the input. For example, in

  from e in emps
    order e.sal
    group e.deptno compute {s = sum over e.sal}

the `sum` function is applied to an `int list`. (The element type,
`int`, is due to the `sal` field; the collection type is `list`
because of the preceding `order` step.) But `sum` has type `int bag ->
int`. Should we allow an aggregate function to be applied to a
collection of different type?

Yes! We don't want to throw away the collection type (most aggregate
functions ignore the order of elements, but it is important for
`FIRST_VALUE` and `STRING_AGG`) but neither do we wish to force people
to write a `list` and `bag` version of each user-defined aggregate
function. Therefore the rules are as follows:

 * We allow both `list` and `bag` aggregate functions;
 * Any aggregate function can be applied to both `list` and `bag` input
   (inserting a list-to-bag or bag-to-list converter as necessary);
 * If both `list` and `bag` forms exist, choose the one that matches
   the input collection type.

These rules apply to `over` in the `compute` clause of a `group` step,
to `over` in a `compute` step, and to an `into` step.

The `elements` keyword (corresponding to the output of the step
preceding the `group` or `compute` step) now has appropriate
collection type.

Remove the overloads of built-in aggregate functions `count`, `max`,
`min`, `empty`, `nonEmpty`; their argument is now `bag` (which can be
created from `list` at zero cost). `only` remains overloaded because
we would like to use it as a function in queries.

So that this feature can be tested against genuine built-in functions,
add several built-in functions in the `Test` structure: `bagSum`,
`listSum`, `overSum`, etc.  The `excludeStructures` property (added in

Wrappers are added during translation from AST to Core. `Resolver`
stores the `ordered` parameter and adds a converter to the argument of
aggregate function when its collection kind differs from the input.

Fixes #271