[MOREL-305] The "cannot derive label for compute expression" error currently causes a crash
Before this change, the "cannot derive label for compute
expression" error is an instance of `class
IllegalArgumentException`, which means that the error
causes the program (such as the interactive shell) to abort.
After this change, it is an instance of `class
CompileException`, which can be handled, and also contains
position information.

Previously, implicit labels were assigned by the parser, as
the AST is being built. If an implicit label could not be
derived, the parser would abort, and would not consume the
rest of the input. Now, implicit labels are assigned during
validation/type derivation. Methods such as
`TypeResolver.deduceType` return a copy of the `Ast.Record`
node with the labels made explicit.

Some tests that checked for implicit labels and duplicate
labels used to be parser tests and have been converted to
validation tests.

Fixes #305