[MOREL-310] Add validation-mode to scripts, with syntax `:t`, to check the type of an expression without evaluating
Fixes #310

Add syntax to test scripts to validate expressions/declarations
and display their type.

If you write

  1;
  > val it = 1 : int

it is executed, but if you prefix with ':t', like this:

  :t 1;
  > val it : int

it only its type is derived.

Note that ':t' can occur on its own line, which is useful if the
statement is intended to produce warnings with positions. And it
can occur in the middle of a statement. (You should only do this
if the statement is preceded by a comment.)