[MOREL-309] Add an `--echo` flag to script runner
Fixes #309

Add an `--echo` flag to `Script.main`. The `--echo` option
causes output to be printed to both the log file and stdout,
and makes it easier to find sections of a script that are
slow or hang.

Factor a new `class Script` out of `ScriptTest`. For each
`.smli` file, `ScriptTest` generates a test, and the test
runs by instantating and invoking a `Script`. Move the
`ScriptTest.main()` method to `Script`, so this is now the
way to invoke a script from the command line.

In #300 we added the `morel` launcher shell script. But FYI
you can also launch scripts directly using Maven, like this:

  ./mvnw -q test-compile exec:java \
      -Dexec.mainClass="net.hydromatic.morel.Script" \
      -Dexec.classpathScope=test \
      -Dexec.args="--echo script/wordle.smli