Add signatures for standard library, and verify that built-in types match
Building on signature parsing (added in #315), this change
adds `lib/bool.sig` for the `BOOLEAN` signature that defines
the `Boolean` structure, `lib/listpair.sig` for the
`LIST_PAIR` signature that defines the `ListPair` structure,
and so forth.

We also add `class SignatureChecker`, invoked from `ListTest`
to ensure the functions, exceptions, datatypes, and their
constructors, that are in the signatures are matched by
definitions in `enum BuiltIn` and its inner classes.

After comparing built-ins with the signature files we noticed
several inconsistencies, which we fix with the changes below.

Add Bool structure, with methods `fromString`, `toString`,
`not`.

In BuiltIn:
 * add built-in type "exn".
 * in Char, String and Real structures rename `op <` to `<`,
   etc.
 * rename FN_OP_O to FN_O
 * rename GENERAL_OP_O to GENERAL_O;
 * add GENERAL_BEFORE, GENERAL_EXN_MESSAGE, GENERAL_EXN_NAME;
 * remove LIST_OP_AT (which duplicates LIST_AT);
 * ditto BAG_OP_AT.

Add optional description field to `BuiltInExn` for detailed
error messages (e.g., "Subscript [subscript out of bounds]").