The `intersect` and `except` steps should count, and preserve order (#321)
The `intersect` and `except` steps should count the number
of occurrences of each element. When working on a `list`
(but not `bag`) they must preserve order.

The query
  from i in [1,2,3,4,5,1,5,1,2] except [2,4];
currently returns
  > val it = [1,1,1,2,3,5,5] : int list
but should return
  > val it = [1,3,5,1,5,1,2] : int list

The 2 on the right side knocks out the first 2 on the left side,
and subsequent 2s are emitted. The output has the same order as
the input list.

This changes fixes the functions in the `List` and `Bag`
structures and also the query operators.

Fixes #321