Implement queries with unbounded variables by inverting predicates (#217)
Consider the query

  from s where String.isPrefix p "abcd"

The variable `p` is unbounded because not iterating over a
collection. One interpretation is that it is iterating over all values
of its datatype but of course `string` has an infinite number of
values. (Data types like `int` and `real` have a finite number of
values, but too many to practically iterate over. The only
*practically finite* types are `unit`, `boolean`, `char`, sum types
with no payload values or payloads of practically finite types, and
tuples/records whose fields are practically finite types.)

This query can be implemented because the predicate `String.isPrefix p
"abcd"` can be inverted. We know an algorithm that generates the set
of `string` values `p` for which the predicate evaluates to true.
`String.isPrefix p s` inverts to `List.tabulate(String.size s + 1, fn
i => String.substring(s, 0, i))`

Also, importantly, the predicate is in a position in the query where
any value of `p` emited from the `from` clause will be killed if the
predicate does not evaluate to true. If we added `orelse length(p) >
20` to the `where` clause, inverting the `isPrefix` predicate would no
longer be useful.

The definition of predicate inversion requires an *efficient* process
- more efficient than simply trying every possible value. A
*generator* (which is the expression created by inverting a predicate,
subject to a number of other constraints) may declare that it
generates false positives and/or duplicate values. If so, we can run
the values through the predicate and through a `distinct` operator.

In this change, we define several predicate inversion patterns, and
add a framework (`class Expander`, `interface Generator` and `class
Generators`) where new patterns can be added in the future. The
patterns we implement are sufficient for an implementation of Datalog
(arriving shortly in #323).

When applying a rule to invert a predicate, we have a goal pattern
(which may be composite, such as the tuple `(x, y)`) and a list of
predicates. Changing the goal makes it a different request. For
example, we earlier inverted `String.isPrefix p s` with the goal `p`
and the predicate `s = "abcd"`, but if the goal is `s` then the
request is not satisfiable.

As we step through a query, the predicates accumulate. We cannot
invert 'where x > 5' or 'where x < 10' on their own, but after we have
stepped over 'where x > 5' can attempt to invert 'where x < 10'
knowing that 'x > 5' already. And this can be converted to
`List.tabulate (4, fn i => i + 6)`, which yields `[6,7,8,9]`.  (It is
not sufficient for predicates to accumulate; we also need to be in a
location of the query where killing values can propagate to upstream
locations where the variable was read.)

Regarding code organization. We add `class Expander` and `interface
Generator`.  `class SuchThatShuttle` still exists, but just dispatches
to `Expander`.  The role of `Extent` and `Extents` is reduced - they
used to do predicate inversion, but now just deal with expressions
that are over a type or a set of ranges of that type - and we remove
`ExtentTest` entirely.

We have generators for `p elem collection`, `andalso`, `exists`,
recursive functions that express transitive closure, `case` (renaming
and destructuring patterns, as happens to function arguments),
composite patterns (e.g.  `from x, y where (x, y) elem
[(1,true),(2,false)]` and `from x, y where {y, x} elem [{y=1,x=true},
{y=2,x=false}]`).

Other code changes:
 * Add method `Type.elementType()`; refactor uses of `Type.arg(0)` to
   call `elementType()` instead.
 * Add method `Core.Exp.isBoolLiteral(bool)`.
 * Add method `Core.Exp.isExtent()`.
 * Semi-naive evaluation inf `Relational.iterate`. (We should have
   variants of that method that emit `list` and `bag`, and that use
   different incremental execution strategies.)
 * Migrate `MainTest.testFromSuchThat*` tests to `such-that.smli`.
 * Fix off-by-one dates in `scott.smli`.
 * Consolidate `toPat` methods into `CoreBuilder`.
 * Add method `CoreBuilder.simplifyTuple`, simplifying "(#1 x, #2 x)"
   to "x".
 * In `Replacer`, change the map key type from `Id` to `NamedPat`.

Add an example that solves 'factorial' in the Datalog style (not
working yet).

Fixes #217